pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "voxelator"
version="1.0.0"
dependencies = [
    "jsonargparse", "tqdm", "rich",
    "numpy", "matplotlib", "numpy-stl",
    "scikit-image", "open3d",
    "jupyter"
] # project  dependencies


[project.scripts]
totem="voxelator.voxel2stl:voxel2stl_cli"


[tool.setuptools]
packages = ["voxelator"]